# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: BiasharaOS API — POS and business management for Kenyan kiosk owners
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all staff users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a staff user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a staff user's role, status, name or password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
    delete:
      operationId: deactivateUser
      tags: [users]
      summary: Deactivate a staff user (soft delete — preserves sale history)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deactivated

  # ── SETTINGS ──────────────────────────────────────────────────────────
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List all sales
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            description: Filter by date (YYYY-MM-DD)
        - name: limit
          in: query
          required: false
          schema:
            type: number
        - name: offset
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: Sales list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Record a sale (atomic — validates stock and updates it, records debt if on credit)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "409":
          description: Not enough stock for one or more items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get a single sale with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "404":
          description: Not found

  # ── CUSTOMERS ──────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── USERS ──────────────────────────────────────────────────────────────
    AppUser:
      type: object
      required: [id, firstName, lastName, email, role, isActive, createdAt]
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, cashier]
        isActive:
          type: boolean
        profileImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CreateUserInput:
      type: object
      required: [firstName, lastName, email, password]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [admin, cashier]
          default: cashier
    UpdateUserInput:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        role:
          type: string
          enum: [admin, cashier]
        isActive:
          type: boolean
        password:
          type: string
          minLength: 6

    # ── SETTINGS ──────────────────────────────────────────────────────────
...
    SaleItem:
      type: object
      properties:
        productId:
          type: number
        name:
          type: string
        qty:
          type: number
        price:
          type: number
      required: [productId, name, qty, price]
    Sale:
      type: object
      properties:
        id:
          type: number
        receiptNumber:
          type: string
        total:
          type: number
        method:
          type: string
          enum: [Cash, M-PESA, Credit]
        onCredit:
          type: boolean
        customerId:
          type: ["number", "null"]
        customerName:
          type: ["string", "null"]
        mpesaRef:
          type: ["string", "null"]
        mpesaReceipt:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        createdAt:
          type: string
      required: [id, receiptNumber, total, method, onCredit, items, createdAt]
    SaleInput:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        total:
          type: number
        method:
          type: string
          enum: [Cash, M-PESA, Credit]
        onCredit:
          type: boolean
        customerId:
          type: ["number", "null"]
        customerName:
          type: ["string", "null"]
        mpesaRef:
          type: ["string", "null"]
        mpesaReceipt:
          type: ["string", "null"]
      required: [items, total, method]
...